refactor: create http package
