+++ src/main/java/controller/Controller.java
@@ -2,7 +2,7 @@ package controller;
 
 import java.io.OutputStream;
 
-import webserver.HttpRequest;
+import http.HttpRequest;
 
 public interface Controller {
 	void service(HttpRequest httpRequest, OutputStream out);
