+++ src/main/java/http/HttpResponseGenerator.java
@@ -1,4 +1,4 @@
-package utils;
+package http;
 
 import java.io.IOException;
 import java.nio.file.Files;
