+++ src/main/java/http/RequestBody.java
@@ -1,4 +1,4 @@
-package webserver;
+package http;
 
 import java.util.Collections;
 import java.util.Map;
@@ -8,10 +8,6 @@ public class RequestBody {
 
 	public RequestBody(final Map<String, String> body) {
 		this.body = body;
-	}
-
-	public String getBodyElement(String key) {
-		return body.get(key);
 	}
 
 	public Map<String, String> getBody() {
