+++ src/main/java/http/method/HttpMethod.java
@@ -1,4 +1,4 @@
-package method;
+package http.method;
 
 public enum HttpMethod {
 	GET("GET"), POST("POST");
