+++ src/main/java/webserver/RequestHandler.java
@@ -5,10 +5,11 @@ import java.net.Socket;
 import java.nio.charset.StandardCharsets;
 
 import controller.Controller;
+import http.HttpRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import utils.ControllerGenerator;
-import utils.HttpRequestReader;
+import http.HttpRequestReader;
 
 public class RequestHandler implements Runnable {
 	private static final Logger logger = LoggerFactory.getLogger(RequestHandler.class);
